refactor : refactor code about login
