+++ src/main/java/com/reborn/back/login/auth/service/JpaUserDetailsManager.java
@@ -1,12 +1,12 @@
 package com.reborn.back.login.auth.service;
 
 
-import com.reborn.back.domain.user.SocialAccount;
+import com.reborn.back.domain.user.OAuth;
 import com.reborn.back.domain.user.User;
 import com.reborn.back.global.api.ErrorCode;
 import com.reborn.back.global.exception.GeneralException;
 import com.reborn.back.login.auth.jwt.CustomUserDetails;
-import com.reborn.back.login.repository.SocialAccountRepository;
+import com.reborn.back.login.repository.OAuthRepository;
 import com.reborn.back.login.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -20,15 +20,15 @@ import org.springframework.stereotype.Service;
 public class JpaUserDetailsManager implements UserDetailsManager {
     // UserDetailsManager의 구현체로 만들면, Spring Security Filter에서 사용자 정보 회수에 활요할 수 있음
     private final UserRepository userRepository;
-    private final SocialAccountRepository socialAccountRepository;
+    private final OAuthRepository oAuthRepository;
 
     public JpaUserDetailsManager(
             UserRepository userRepository,
             PasswordEncoder passwordEncoder,
-            SocialAccountRepository socialAccountRepository
+            OAuthRepository oAuthRepository
     ) {
         this.userRepository = userRepository;
-        this.socialAccountRepository = socialAccountRepository;
+        this.oAuthRepository = oAuthRepository;
     }
 
     // 실제로 Spring Security 내부에서 사용하는 반드시 구현해야 정상동작을 기대할 수 있는 메소드
@@ -41,27 +41,23 @@ public class JpaUserDetailsManager implements UserDetailsManager {
                     log.warn("유저 정보 없음: {}", username);
                     return new GeneralException(ErrorCode.USER_NOT_FOUND);
                 });
-        SocialAccount socialAccount = (SocialAccount) socialAccountRepository.findByUser(user)
+        OAuth oAuth = (OAuth) oAuthRepository.findByUser(user)
                 .orElseThrow(() -> new GeneralException(ErrorCode.USER_NOT_FOUND));
-        return CustomUserDetails.fromEntity(user, socialAccount);
+        return CustomUserDetails.fromEntity(user, oAuth);
     }
 
     @Override
     // 새로운 사용자 생성
     public void createUser(UserDetails user) {
         log.info("사용자 생성 시도 중: {}", user.getUsername());
-        if (userExists(user.getUsername())) { // 사용자 존재 여부 확인.
-            log.warn("사용자 이미 존재: {}", user.getUsername());
-            throw new GeneralException(ErrorCode.ALREADY_USED_NICKNAME); // 닉네임 중복 예외 발생.
-        }
         try {
             User newUser = userRepository.save(((CustomUserDetails) user).toEntity()); // 사용자 정보 저장.
             CustomUserDetails details = (CustomUserDetails) user;
-            SocialAccount newSocialAccount = new SocialAccount();
+            OAuth newSocialAccount = new OAuth();
             newSocialAccount.setProvider(details.getProvider());
             newSocialAccount.setProviderUserId(details.getUsername());
             newSocialAccount.setUser(newUser);
-            socialAccountRepository.save(newSocialAccount);
+            oAuthRepository.save(newSocialAccount);
             log.info("사용자 생성: {}", user.getUsername());
         } catch (ClassCastException e) {
             log.error("UserDetails ->CustomUserDetails 변환 실패: {}", user.getUsername(), e);
