+++ src/main/java/com/reborn/back/login/auth/utils/OAuth2UserServiceImpl.java
@@ -73,30 +73,3 @@ public class OAuth2UserServiceImpl extends DefaultOAuth2UserService {
         return attributes;
     }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
