+++ src/main/java/com/reborn/back/login/dto/UserResponseDto.java
@@ -36,8 +36,5 @@ public class UserResponseDto {
 
         @Schema(description = "프로필 사진")
         private String profileImage;
-
-        @Schema(description = "닉네임")
-        private String nickname;
     }
 }
