+++ src/main/java/com/reborn/back/login/mapper/UserConverter.java
@@ -1,6 +1,7 @@
 package com.reborn.back.login.mapper;
 
 import com.reborn.back.domain.user.User;
+import com.reborn.back.domain.user.UserInfo;
 import com.reborn.back.login.auth.dto.JwtDto;
 import com.reborn.back.login.dto.UserRequestDto;
 import com.reborn.back.login.dto.UserResponseDto;
@@ -9,16 +10,11 @@ import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
 public class UserConverter {
-    public static User saveUser(UserRequestDto.UserReqDto userReqDto) {
+    public static User saveUser(UserRequestDto userReqDto) {
         return User.builder()
                 .email(userReqDto.getEmail())
-                .username(userReqDto.getUsername())
-                .provider(userReqDto.getProvider())
-                .nickname(userReqDto.getNickname())
+                .name(userReqDto.getUsername())
                 .deviceToken(userReqDto.getDeviceToken())
-                .contentPetId(null)
-                .profileImage(null)
-                .backgroundImage(null)
                 .build();
     }
 
@@ -34,14 +30,12 @@ public class UserConverter {
         return UserResponseDto.UserInfoResDto.builder()
                 .since(user.getCreatedAt())
                 .email(user.getEmail())
-                .nickname(user.getNickname())
                 .build();
     }
 
-    public static UserResponseDto.MainInfoResDto mainDto(User user) {
+    public static UserResponseDto.MainInfoResDto mainDto(UserInfo userInfo) {
         return UserResponseDto.MainInfoResDto.builder()
-                .profileImage(user.getProfileImage())
-                .nickname(user.getNickname())
+                .profileImage(userInfo.getImg())
                 .build();
     }
 }
