--- src/main/java/com/reborn/back/login/repository/UuidRepository.java
@@ -1,8 +1,0 @@
-package com.reborn.back.login.repository;
-
-import com.reborn.back.domain.entity.Uuid;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UuidRepository extends JpaRepository<Uuid, Long> {
-
-}
