--- src/main/java/com/reborn/back/mypage/dto/MessageDto.java
@@ -1,15 +1,0 @@
-package com.reborn.back.mypage.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Builder
-public class MessageDto {
-    private String to;
-//    String content;
-}
