--- src/main/java/com/reborn/back/mypage/dto/SmsRequestDto.java
@@ -1,21 +1,0 @@
-package com.reborn.back.mypage.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Builder
-public class SmsRequestDto {
-    private String type;
-    private String contentType;
-    private String countryCode;
-    private String from;
-    private String content;
-    private List<MessageDto> messages;
-}
