--- src/main/java/com/reborn/back/mypage/dto/SmsResponseDto.java
@@ -1,20 +1,0 @@
-package com.reborn.back.mypage.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Builder
-public class SmsResponseDto {
-    private String requestId;
-    private LocalDateTime requestTime;
-    private String statusCode;
-    private String statusName;
-    private String smsConfirmNum;
-}
